requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart